Worker clears locks on shutdown [PLAT-1593]
In some cases the ClearLocks plugin from Delayed::Job may fail to clear
the locked_at/locked_by fields on a job because it was interrupted at a
bad time (see [PLAT-1590]). The delayed_worker is still removed since
the heartbeat is using a separate connection, leaving orphaned jobs
behind that are not re-run until the 4 hour max run time has passed.

This change will require the worker to clear its own locks when it is
shutting down before it removes itself from the delayed_workers table.